API updated to minimal hosting model (Program.cs/Startup.cs)
